chore!: Update for changes to `sn_node` CLI
The `sn_node` CLI is changing to make `--first` a flag, rather than
taking a value for the local/public address. The `--ip` argument is now
used to set the `--local-addr` for `sn_node` for all nodes, not just
genesis.

BREAKING CHANGE: For genesis nodes, `sn_node` will now be invoked with
`--first` without a value, which is not compatible with older versions
of `sn_node`.