feat: introduce skip auto port forwarding flag
This option is used to pass `--skip-auto-port-forwarding` to the node binary. Using the `--local`
flag also has the same effect, but it doesn't make sense to always use the `--local` flag, because
port forwarding can also be disabled for connecting to remote networks. Providing a separate option
avoids any potential confusion.